refactor: 💡 contract views refactor first level
