+++ packages/frontend/src/App.jsx
@@ -15,7 +15,7 @@ import { LOGIN_SUCCESS } from './redux/auth/constants';
 import AuthorsNewArticle from './components/ui/author/AuthorsnewArticle/AuthorsNewArticle';
 import AllPublications from './components/ui/author/allPublications/AllPublications/AllPublications';
 import TopBar from './utils/components/TopBar';
-import GeneralContractView from './components/SalesScreens/ContractViews/GeneralContractView';
+import GeneralContractView from './components/ui/salesman/contractViews/GeneralContractView/GeneralContractView';
 import Snackbar from './components/ui/Snackbar/Snackbar.jsx';
 import Sales from './components/ui/salesman/Sales/Sales';
 import DeadArticle from './components/ui/author/DeadArticle/DeadArticle';
